Implemented part of the BUY logic. Tested in Main class.

+++ src/order/Crypto.java
@@ -4,10 +4,10 @@ import java.math.BigDecimal;
 
 public class Crypto {
     public int supply;
-    public BigDecimal price;
+    public double price;
     public String symbol;
 
-    public Crypto(String symbol, BigDecimal price, int supply) { // Constructor. Will be used to create various coins if needed.
+    public Crypto(String symbol, double price, int supply) { // Constructor. Will be used to create various coins if needed.
         this.symbol = symbol;
         this.price = price;
         this.supply = supply;
